# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.Product:
    properties:
      brand:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      original_price:
        type: integer
      price:
        type: integer
      product_url:
        type: string
      sku:
        type: string
      updated_at:
        type: string
    type: object
  domain.ProductList:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      products:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      total:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.rateResponse:
    properties:
      rate:
        type: number
    type: object
info:
  contact: {}
  description: Product catalog API for store77.net marketplace
  title: Store Marketplace API
  version: "1.0"
paths:
  /brands:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List all brands
      tags:
      - products
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List all categories
      tags:
      - categories
  /categories/{id}:
    get:
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get category by ID
      tags:
      - categories
  /exchange/rate:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.rateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get USDT/RUB exchange rate
      tags:
      - exchange
  /products:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 24
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort (e.g. price:asc,name:desc)
        in: query
        name: sort_fields
        type: string
      - description: Category UUID
        in: query
        name: category_id
        type: string
      - description: Brand filter
        in: query
        name: brand
        type: string
      - description: Min price (RUB)
        in: query
        name: min_price
        type: integer
      - description: Max price (RUB)
        in: query
        name: max_price
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProductList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List products
      tags:
      - products
  /products/{id}:
    get:
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get product by ID
      tags:
      - products
swagger: "2.0"